[WIP] Resource page ergonomics - footer with actions
