+++ src/app/js/public/resource/Detail.js
@@ -1,54 +1,63 @@
 import React, { PropTypes } from 'react';
 import { connect } from 'react-redux';
-import { CardHeader, CardActions } from 'material-ui/Card';
-import { Link } from 'react-router';
-import FlatButton from 'material-ui/FlatButton';
-import compose from 'recompose/compose';
-import translate from 'redux-polyglot/translate';
+import { CardText } from 'material-ui/Card';
+import memoize from 'lodash.memoize';
 
-import { fromResource } from '../selectors';
-import Card from '../../lib/Card';
-import { polyglot as polyglotPropTypes } from '../../propTypes';
-import { isLoggedIn } from '../../user';
-import DetailProperties from './DetailProperties';
+import {
+    fromResource,
+    fromPublication,
+} from '../selectors';
+import Property from '../Property';
 
-export const DetailComponent = ({ resource, isLogged, p: polyglot }) => (
-    <Card className="detail">
-        <CardHeader title={'Properties'} />
-        <DetailProperties />
-        <CardActions>
-            {
-                (isLogged ? ['edit', 'hide', 'add-field'] : ['add-field'])
-                .map(name => (
-                    <Link key={name} to={{ pathname: `/resource/${name}`, query: { uri: resource.uri } }}>
-                        <FlatButton className={`${name}-resource`} label={polyglot.t(name)} primary />
-                    </Link>
-                ))
-            }
-        </CardActions>
-    </Card>
+const styles = {
+    container: {
+        display: 'flex',
+        flexDirection: 'column',
+    },
+    item: memoize((index, total) => ({
+        borderBottom: index < total - 1 ? '1px solid rgb(224, 224, 224)' : 'none',
+        paddingBottom: index < total - 1 ? '3rem' : 0,
+        paddingTop: '2rem',
+    })),
+};
+
+export const DetailPropertiesComponent = ({ resource, collectionFields, documentFields }) => (
+    <CardText className="detail-properties" style={styles.container}>
+        {collectionFields.map((field, index) => (
+            <Property
+                key={field.name}
+                resource={resource}
+                field={field}
+                style={styles.item(index, collectionFields.length)}
+            />
+        ))}
+        {documentFields.filter(({ name }) => !!resource[name]).map((field, index) => (
+            <Property
+                key={field.name}
+                resource={resource}
+                field={field}
+                style={styles.item(index, documentFields.length)}
+            />
+        ))}
+    </CardText>
 );
 
-DetailComponent.defaultProps = {
+DetailPropertiesComponent.defaultProps = {
     resource: null,
 };
 
-DetailComponent.propTypes = {
-    resource: PropTypes.shape({
-        uri: PropTypes.string.isRequired,
-    }),
-    isLogged: PropTypes.bool.isRequired,
-    p: polyglotPropTypes.isRequired,
+DetailPropertiesComponent.propTypes = {
+    resource: PropTypes.shape({}),
+    collectionFields: PropTypes.arrayOf(PropTypes.object).isRequired,
+    documentFields: PropTypes.arrayOf(PropTypes.object).isRequired,
 };
 
 const mapStateToProps = state => ({
     resource: fromResource.getResourceLastVersion(state),
-    isLogged: isLoggedIn(state),
+    collectionFields: fromPublication.getRootCollectionFields(state),
+    documentFields: fromPublication.getDocumentFields(state),
 });
 
 const mapDispatchToProps = {};
 
-export default compose(
-    translate,
-    connect(mapStateToProps, mapDispatchToProps),
-)(DetailComponent);
+export default connect(mapStateToProps, mapDispatchToProps)(DetailPropertiesComponent);
