+++ src/app/js/public/resource/sagas/addFieldToResource.spec.js
@@ -51,9 +51,7 @@ describe('handleAddFieldToResource', () => {
         saga.next();
         saga.next({ reason: 'reason' });
         saga.next('request');
-        let next = saga.next({ response: 'response' });
+        const next = saga.next({ response: 'response' });
         expect(next.value).toEqual(put(addFieldToResourceSuccess('response')));
-        next = saga.next();
-        expect(next.value).toEqual(put(push({ pathname: '/resource', query: { uri: 'uri' } })));
     });
 });
